Rough logic for some player actions
Put down what I figure a good start for the player command code should
be. It shouldn't compile and still needs more modifications to the class
structure, but it is a start. Also wanted a chance to make sure I was
using Github correctly.

Wasn't sure exactly how the items/inventory work so I didn't touch that
stuff too much.