Fleshed out Rooms and added a little bit for Items
I tried to help create a more solid idea how of the rooms are structured
and simplify them a bit. I left in the ConnectingRooms, but I don't
think they will necessarily be needed. Also tried to add a bit to Items,
but I am still not sure on how they are being handled really.